requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
cloudinary==1.36.0
reportlab==4.0.4
requests==2.31.0
gunicorn==21.2.0